+++ Test/Graphics/dx9_dot3_bump_mapping/dx9_dot3_bump_mapping.cpp
@@ -223,7 +223,7 @@ int WINAPI WinMain( HINSTANCE hInstance,
 
     UnregisterClass( "MY_WINDOWS_CLASS", winClass.hInstance );
 
-    return uMsg.wParam;
+    return (int)uMsg.wParam;
 }
 
 //-----------------------------------------------------------------------------
